+++ packages/cashc/src/ast/AST.ts
@@ -165,6 +165,19 @@ export class BranchNode extends StatementNode {
   }
 }
 
+export class DoWhileNode extends StatementNode {
+  constructor(
+    public condition: ExpressionNode,
+    public block: BlockNode,
+  ) {
+    super();
+  }
+
+  accept<T>(visitor: AstVisitor<T>): T {
+    return visitor.visitDoWhile(this);
+  }
+}
+
 export class BlockNode extends Node {
   symbolTable?: SymbolTable;