+++ packages/cashc/src/print/OutputSourceCodeTraversal.ts
@@ -29,6 +29,7 @@ import {
   ConsoleStatementNode,
   ConsoleParameterNode,
   SliceNode,
+  DoWhileNode,
 } from '../ast/AST.js';
 import AstTraversal from '../ast/AstTraversal.js';
 
@@ -160,6 +161,15 @@ export default class OutputSourceCodeTraversal extends AstTraversal {
 
     this.addOutput('\n');
 
+    return node;
+  }
+
+  visitDoWhile(node: DoWhileNode): Node {
+    this.addOutput('do ', true);
+    node.block = this.visit(node.block) as BlockNode;
+    this.addOutput(' while (', true);
+    node.condition = this.visit(node.condition);
+    this.addOutput(');\n');
     return node;
   }