+++ website/docs/releases/release-notes.md
@@ -4,6 +4,10 @@ title: Release Notes
 
 ## v0.13.0-next.0
 
+#### cashc compiler
+- :sparkles: Add support for `do {} while ()` loops.
+- :racehorse: Add optimisations for negated number comparisons.
+
 #### CashScript SDK
 
 - :hammer_and_wrench: Update default VM target to `BCH_2026_05`.