+++ website/docs/releases/release-notes.md
@@ -2,6 +2,16 @@
 title: Release Notes
 ---
 
+## v0.13.0-next.0
+
+#### cashc compiler
+- :sparkles: Add support for `do {} while ()` loops.
+- :racehorse: Add optimisations for negated number comparisons.
+
+#### CashScript SDK
+
+- :hammer_and_wrench: Update default VM target to `BCH_2026_05`.
+
 ## v0.12.0
 
 #### CashScript SDK